cli: move InitCommand away from lifecycle
