+++ rockcraft/commands/lifecycle.py
@@ -149,42 +149,3 @@ class PackCommand(_LifecycleCommand):
         all parts.
         """
     )
-
-
-class InitCommand(_LifecycleCommand):
-    """Initialize a rockcraft project."""
-
-    name = "init"
-    help_msg = "Initialize a rockcraft project"
-    overview = textwrap.dedent(
-        """
-        Initialize a rockcraft project by creating a minimalist,
-        yet functional, rockcraft.yaml file in the current directory.
-        """
-    )
-
-    _INIT_TEMPLATE_YAML = textwrap.dedent(
-        """\
-            name: my-rock-name # the name of your ROCK
-            base: ubuntu:22.04 # the base environment for this ROCK
-            version: '0.1' # just for humans. Semantic versioning is recommended
-            summary: Single-line elevator pitch for your amazing ROCK # 79 char long summary
-            description: |
-                This is my my-rock-name's description. You have a paragraph or two to tell the
-                most important story about it. Keep it under 100 words though,
-                we live in tweetspace and your description wants to look good in the
-                container registries out there.
-            license: GPL-3.0 # your application's SPDX license
-            platforms: # The platforms this ROCK should be built on and run on
-                amd64:
-
-            parts:
-                my-part:
-                    plugin: nil
-            """
-    )
-
-    @overrides
-    def run(self, parsed_args):
-        """Run the command."""
-        lifecycle.init(self._INIT_TEMPLATE_YAML)
